+++ tools/uart-flash-server/ufserver.c
@@ -88,7 +88,7 @@ void printmsg(const char *msg)
 }
 
 
-int hal_flash_write(uint32_t address, const uint8_t *data, int len)
+int hal_flash_write(uintptr_t address, const uint8_t *data, int len)
 {
     (void)address;
     (void)data;
@@ -96,7 +96,7 @@ int hal_flash_write(uint32_t address, const uint8_t *data, int len)
     return -1;
 }
 
-int hal_flash_erase(uint32_t address, int len)
+int hal_flash_erase(uintptr_t address, int len)
 {
     (void)address;
     (void)len;
